Eslint, use src, update deps, use a utils module
