+++ .eslintignore
@@ -1,0 +1,4 @@
+test
+dist
+node_modules
+example
