--- .jshintrc
@@ -1,24 +1,0 @@
-{
-  "esversion": 11,
-  "curly": true,
-  "eqeqeq": true,
-  "newcap": true,
-  "noarg": true,
-  "noempty": true,
-  "nonew": true,
-  "sub": true,
-  "undef": true,
-  "unused": true,
-  "trailing": true,
-  "boss": true,
-  "eqnull": true,
-  "strict": true,
-  "immed": true,
-  "expr": true,
-  "latedef": "nofunc",
-  "quotmark": "single",
-  "validthis": true,
-  "indent": 2,
-  "node": true,
-  "browser": true
-}
