+++ test/public-api.js
@@ -1,7 +1,7 @@
-'use strict';
 
-import test from "./lib/base.js";
-import dragula from "../dragula.js";
+
+import test from './lib/base.js';
+import {dragula} from '../src/dragula.js';
 
 test('public api matches expectation', function (t) {
   t.equal(typeof dragula, 'function', 'dragula is a function');
